split utils into different files, update scripts
